UI Additions
Hey Will. I did some minor work to the UI so hopefully it will inspire you to add WCEasySettings so that we can add some nice features to this emulator. I used most of the same parameters from your iNDS project to build the UI. I can link the objects from the main storyboard to code once we get the settings up and running.